Refactor `Command`
