+++ src/main/java/spleetwaise/address/logic/Logic.java
@@ -4,13 +4,13 @@ import java.nio.file.Path;
 
 import javafx.collections.ObservableList;
 import spleetwaise.address.commons.core.GuiSettings;
-import spleetwaise.address.logic.commands.CommandResult;
-import spleetwaise.address.logic.commands.exceptions.CommandException;
 import spleetwaise.address.logic.parser.exceptions.ParseException;
 import spleetwaise.address.model.AddressBookModel;
 import spleetwaise.address.model.ReadOnlyAddressBook;
 import spleetwaise.address.model.person.Person;
 import spleetwaise.commons.exceptions.SpleetWaiseCommandException;
+import spleetwaise.commons.logic.commands.CommandResult;
+import spleetwaise.commons.logic.commands.exceptions.CommandException;
 import spleetwaise.transaction.model.transaction.Transaction;
 
 /**
