+++ src/main/java/spleetwaise/address/logic/commands/AddCommand.java
@@ -10,9 +10,11 @@ import static spleetwaise.address.logic.parser.CliSyntax.PREFIX_TAG;
 
 import spleetwaise.address.commons.util.ToStringBuilder;
 import spleetwaise.address.logic.Messages;
-import spleetwaise.address.logic.commands.exceptions.CommandException;
 import spleetwaise.address.model.person.Person;
-import spleetwaise.commons.CommonModel;
+import spleetwaise.commons.logic.commands.Command;
+import spleetwaise.commons.logic.commands.CommandResult;
+import spleetwaise.commons.logic.commands.exceptions.CommandException;
+import spleetwaise.commons.model.CommonModel;
 
 /**
  * Adds a person to the address book.
