--- src/main/java/spleetwaise/address/logic/commands/Command.java
@@ -1,18 +1,0 @@
-package spleetwaise.address.logic.commands;
-
-import spleetwaise.address.logic.commands.exceptions.CommandException;
-
-/**
- * Represents a command with hidden internal logic and the ability to be executed.
- */
-public abstract class Command {
-
-    /**
-     * Executes the command and returns the result message.
-     *
-     * @return feedback message of the operation result for display
-     * @throws CommandException If an error occurs during command execution.
-     */
-    public abstract CommandResult execute() throws CommandException;
-
-}
