+++ src/main/java/spleetwaise/address/logic/commands/EditCommand.java
@@ -19,7 +19,6 @@ import spleetwaise.address.commons.core.index.Index;
 import spleetwaise.address.commons.util.CollectionUtil;
 import spleetwaise.address.commons.util.ToStringBuilder;
 import spleetwaise.address.logic.Messages;
-import spleetwaise.address.logic.commands.exceptions.CommandException;
 import spleetwaise.address.model.AddressBookModel;
 import spleetwaise.address.model.person.Address;
 import spleetwaise.address.model.person.Email;
@@ -28,7 +27,10 @@ import spleetwaise.address.model.person.Person;
 import spleetwaise.address.model.person.Phone;
 import spleetwaise.address.model.person.Remark;
 import spleetwaise.address.model.tag.Tag;
-import spleetwaise.commons.CommonModel;
+import spleetwaise.commons.logic.commands.Command;
+import spleetwaise.commons.logic.commands.CommandResult;
+import spleetwaise.commons.logic.commands.exceptions.CommandException;
+import spleetwaise.commons.model.CommonModel;
 
 /**
  * Edits the details of an existing person in the address book.
