--- src/main/java/spleetwaise/address/logic/commands/exceptions/CommandException.java
@@ -1,16 +1,0 @@
-package spleetwaise.address.logic.commands.exceptions;
-
-import spleetwaise.commons.exceptions.SpleetWaiseCommandException;
-
-/**
- * Represents an error that occurs during command execution in the address component.
- */
-public class CommandException extends SpleetWaiseCommandException {
-    public CommandException(String message) {
-        super(message);
-    }
-
-    public CommandException(String message, Throwable cause) {
-        super(message, cause);
-    }
-}
