+++ src/main/java/spleetwaise/transaction/logic/commands/AddCommand.java
@@ -7,9 +7,10 @@ import static spleetwaise.transaction.logic.parser.CliSyntax.PREFIX_DATE;
 import static spleetwaise.transaction.logic.parser.CliSyntax.PREFIX_DESCRIPTION;
 
 import spleetwaise.address.commons.util.ToStringBuilder;
-import spleetwaise.address.logic.commands.CommandResult;
-import spleetwaise.commons.CommonModel;
-import spleetwaise.transaction.logic.commands.exceptions.CommandException;
+import spleetwaise.commons.logic.commands.Command;
+import spleetwaise.commons.logic.commands.CommandResult;
+import spleetwaise.commons.logic.commands.exceptions.CommandException;
+import spleetwaise.commons.model.CommonModel;
 import spleetwaise.transaction.model.transaction.Transaction;
 
 /**
