+++ src/main/java/spleetwaise/transaction/logic/parser/Parser.java
@@ -1,7 +1,7 @@
 package spleetwaise.transaction.logic.parser;
 
 import spleetwaise.address.logic.parser.exceptions.ParseException;
-import spleetwaise.transaction.logic.commands.Command;
+import spleetwaise.commons.logic.commands.Command;
 
 /**
  * Represents a Parser that is able to parse user input into a {@code Command} of type {@code T}.
