+++ src/test/java/spleetwaise/address/logic/parser/CommandParserTestUtil.java
@@ -2,8 +2,8 @@ package spleetwaise.address.logic.parser;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
 
-import spleetwaise.address.logic.commands.Command;
 import spleetwaise.address.logic.parser.exceptions.ParseException;
+import spleetwaise.commons.logic.commands.Command;
 
 /**
  * Contains helper methods for testing command parsers.
@@ -19,7 +19,7 @@ public class CommandParserTestUtil {
             Command expectedCommand
     ) {
         try {
-            Command command = parser.parse(userInput);
+            spleetwaise.commons.logic.commands.Command command = parser.parse(userInput);
             assertEquals(expectedCommand, command);
         } catch (ParseException pe) {
             throw new IllegalArgumentException("Invalid userInput.", pe);
@@ -30,7 +30,10 @@ public class CommandParserTestUtil {
      * Asserts that the parsing of {@code userInput} by {@code parser} is unsuccessful and the error message equals to
      * {@code expectedMessage}.
      */
-    public static void assertParseFailure(Parser<? extends Command> parser, String userInput, String expectedMessage) {
+    public static void assertParseFailure(
+            Parser<? extends Command> parser, String userInput,
+            String expectedMessage
+    ) {
         try {
             parser.parse(userInput);
             throw new AssertionError("The expected ParseException was not thrown.");
