+++ src/test/java/spleetwaise/commons/logic/commands/CommandResultTest.java
@@ -1,4 +1,4 @@
-package spleetwaise.address.logic.commands;
+package spleetwaise.commons.logic.commands;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertFalse;
