+++ src/test/java/spleetwaise/commons/model/CommonModelTest.java
@@ -1,4 +1,4 @@
-package spleetwaise.commons;
+package spleetwaise.commons.model;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertThrows;
